Fixed Role Issues and Added Playbooks
Added the playbook-debian-server and playbook-debian-workstation for
updating my local systems.

Corrected an issue with the MOTD role where hddtemp doesn't exist, but
this has been removed entirely since it's not used with the
update-motd.d scripts.

Updated userprofile to set zsh as the preferred shell for the default
user *if* it's installed.